fix: address review comments for torch.compile and int32/int64 type promotion
1. Split force_topk_sample into _force_topk_process_logits (compiled,
   pure tensor ops Phase 1/2/3/5) and force_topk_sample wrapper
   (uncompiled Phase 4 with generators). Fixes graph break caused by
   torch.compile tracing dict[int, torch.Generator].

2. Replace .long() with .to(torch.int32) in 3 places to avoid NPU
   implicit int32/int64 type promotion:
   - sampler.py: token_ids comparison with token_index (int32)
   - sampler.py: token_ids concatenation with topk_indices (int32)
   - topk_map.py: vocab_ids comparison with token_index (int32)

+++ vllm_ascend/sample/sampler.py
@@ -317,7 +317,7 @@ class AscendSampler(Sampler):
 
         # Single search: find sampled token position in token_index.
         hit = (
-            cdist.token_index == token_ids.long().unsqueeze(-1)
+            cdist.token_index == token_ids.to(torch.int32).unsqueeze(-1)
         )  # [B, k]
         pos = hit.long().argmax(dim=-1)  # [B]
         found = hit.any(dim=-1)  # [B]
@@ -337,7 +337,7 @@ class AscendSampler(Sampler):
         )
 
         indices = torch.cat(
-            (token_ids.unsqueeze(-1), topk_indices), dim=1
+            (token_ids.unsqueeze(-1).to(torch.int32), topk_indices), dim=1
         )
         logprobs_cat = torch.cat(
             (token_logprobs.unsqueeze(-1), topk_logprobs), dim=1
