+++ vllm_ascend/sample/topk_map.py
@@ -49,7 +49,7 @@ class CompactDist:
         Returns:
             [B] float32 tensor of logprob values (or -inf on miss).
         """
-        hit = self.token_index == vocab_ids.long().unsqueeze(-1)
+        hit = self.token_index == vocab_ids.to(torch.int32).unsqueeze(-1)
         pos = hit.long().argmax(dim=-1)  # [B]
         val = self.logprobs.gather(
             1, pos.unsqueeze(1)
